[Refactor] #3 기존 클래스를 다른 패키지로 이동
