+++ src/main/java/com/donet/donet/home/adapter/in/web/MainPageController.java
@@ -1,6 +1,5 @@
-package com.donet.donet;
+package com.donet.donet.home.adapter.in.web;
 
-import com.donet.donet.dto.ResponseMainPageDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -8,8 +7,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("")
-public class Controller {
-
+public class MainPageController {
     @Operation(summary = "메인페이지", description = "메인 페이지 조회 API입니다.")
     @GetMapping("/main")
     public ResponseMainPageDTO mainPage(){
